pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mne-beam-pipeline"
version = "0.1.0"
description = "Config-driven MNE beamforming pipeline scaffold"
requires-python = ">=3.10"
dependencies = [
  "mne>=1.7",
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "h5py>=3.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
